Update default envlist
Usually, I want to run tests and not build docs or perform other
checking, so make that the default and keep the others for expressly
requested. Having `{posargs}` in the non-pytest environments also
means these commands would break if args are passed.

Also, add pypy3-py2neo2 to environment matrix.